+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to this project will be documented in this file. The format is based on
 [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## 2025-12-24
+
+### 🔄 Changed
+
+- Container apps for cache and search builders changed to container app jobs
+
 ## 2025-12-15
 
 ### ✨ Added